Dont snap the panel instantly to the destination when setting a new state during animation
Calling cancel on an Animation triggers the end listener immediately in the same callstack, which in turn
triggers the PanelSlideListener. This potentially leads to confusing behaviour and wrong order of operations
when the consumer uses the PanelSlideListener.

Just changing the destination of the movement and discarding the previous slide request is a more reasonable approach.